change id type in db script & add spring jpa configuration

+++ store/src/main/resources/db-data/init/setup.sql
@@ -7,7 +7,7 @@ drop table if exists users;
 
 create table products
 (
-    id            integer PRIMARY KEY AUTO_INCREMENT,
+    id            bigint PRIMARY KEY AUTO_INCREMENT,
     name          varchar(255),
     last_name     varchar(255),
     price         integer,
@@ -17,15 +17,16 @@ create table products
 );
 create table users
 (
-    id            integer PRIMARY KEY AUTO_INCREMENT,
+    id            bigint PRIMARY KEY AUTO_INCREMENT,
     name          varchar(255),
     email         varchar(255)
 );
 create table users_products
 (
-    id            integer PRIMARY KEY AUTO_INCREMENT,
-    product_id    integer not null,
-    user_id       integer not null,
+    id            bigint PRIMARY KEY AUTO_INCREMENT,
+    product_id    bigint not null,
+    user_id       bigint not null,
     CONSTRAINT FK_Product FOREIGN KEY (product_id) REFERENCES products(id),
     CONSTRAINT FK_User FOREIGN KEY (user_id) REFERENCES users(id)
 );
+
